Split resnet and fastrcnn
